ta-commit2

+++ src/main/java/com/websitedungcuthethao/repository/AnhSanPhamRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.websitedungcuthethao.entity.AnhSanPham;
 
-public interface AnhSanPhamRepository extends JpaRepository<AnhSanPham, Integer>{
+public interface AnhSanPhamRepository extends JpaRepository<AnhSanPham, Long>{
 }
