+++ src/main/java/com/websitedungcuthethao/repository/DanhMucRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.websitedungcuthethao.entity.DanhMuc;
 
-public interface DanhMucRepository extends JpaRepository<DanhMuc, Integer> {
+public interface DanhMucRepository extends JpaRepository<DanhMuc, Long> {
 	DanhMuc findByTen(String ten);
 }
