+++ src/main/java/com/websitedungcuthethao/repository/DiaChiRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.websitedungcuthethao.entity.DiaChi;
 
-public interface DiaChiRepository extends JpaRepository<DiaChi, Integer> {
+public interface DiaChiRepository extends JpaRepository<DiaChi, Long> {
 	
 }
