+++ src/main/java/com/websitedungcuthethao/repository/NguoiDungRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.websitedungcuthethao.entity.NguoiDung;
 
-public interface NguoiDungRepository extends JpaRepository<NguoiDung,Integer> {
+public interface NguoiDungRepository extends JpaRepository<NguoiDung,Long> {
 	NguoiDung findOneByTenDangNhap(String tenDangNhap);
 	List<NguoiDung> findByTenAndHo(String ten,String ho);
