+++ src/main/java/com/websitedungcuthethao/repository/NhaCungCapRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.websitedungcuthethao.entity.NhaCungCap;
 
-public interface NhaCungCapRepository extends JpaRepository<NhaCungCap, Integer> {
+public interface NhaCungCapRepository extends JpaRepository<NhaCungCap, Long> {
 	List<NhaCungCap> findByTenNhaCungCap(String ten);
 }
