+++ src/main/java/com/websitedungcuthethao/repository/ThuocTinhSanPhamRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.websitedungcuthethao.entity.ThuocTinhSanPham;
 
-public interface ThuocTinhSanPhamRepository extends JpaRepository<ThuocTinhSanPham, Integer> {
+public interface ThuocTinhSanPhamRepository extends JpaRepository<ThuocTinhSanPham, Long> {
 
 }
