+++ src/main/java/com/websitedungcuthethao/service/INguoiDungService.java
@@ -8,7 +8,7 @@ public interface INguoiDungService {
 	List<NguoiDung> findAll();
 	void saveNguoiDung(NguoiDung nguoiDung);
 	void UpdateNguoiDung(NguoiDung nguoiDung);
-	NguoiDung findById(int id);
+	NguoiDung findById(Long id);
 	List<NguoiDung> findByTenAndHo(String ten,String ho);
 	
 }
